+++ app/layout.tsx
@@ -6,9 +6,9 @@ import { ColorSchemeScript } from '@mantine/core';
 import { Analytics } from '@vercel/analytics/react';
 import { SpeedInsights } from '@vercel/speed-insights/next';
 import classnames from 'classnames';
-import Footer from '@/components/Common/footer/Footer';
-import Header from '@/components/Common/header/Header';
-import Provider from '@/components/Common/Provider';
+import Footer from '@/components/footer/Footer';
+import Header from '@/components/header/Header';
+import Provider from '@/components/provider/Provider';
 import '@/styles/globals.scss';
 import '@mantine/spotlight/styles.css';
 import styles from './layout.module.scss';
