+++ components/featuredPost/FeaturedPost.module.scss
