+++ components/featuredPost/FeaturedPost.tsx
