+++ components/searchBar/SearchBar.module.scss
