+++ utils/calculateTagPageCount.ts
@@ -1,4 +1,4 @@
-import { POSTS_PER_PAGE } from '@/components/Common/pagination/Pagination';
+import { POSTS_PER_PAGE } from '@/components/pagination/Pagination';
 
 const calculateTagPageCount = (numberOfTags: number) => {
   const pageCount = Math.ceil(numberOfTags / POSTS_PER_PAGE);
